refactor cache: split interface and implementation
