+++ internal/cache/cache.go
@@ -1,9 +1,11 @@
 package cache
 
 import (
+	"context"
 	"errors"
+	"time"
 
-	"github.com/gadelkareem/cachita"
+	"github.com/gphotosuploader/googlemirror/api/photoslibrary/v1"
 )
 
 // ErrCacheMiss is returned when a object is not found in cache.
@@ -13,11 +15,18 @@ type Cache interface {
 	albumsCache
 }
 
-// CachitaCache implements Cache using `gadelkareem/cachita` package.
-type CachitaCache struct {
-	cache cachita.Cache
-}
+// albumsCache is used to store and retrieve previously obtained Albums.
+type albumsCache interface {
+	// GetAlbum returns Album data from the cache corresponding to the specified key.
+	// It will return ErrCacheMiss if there is no cached Album with that key.
+	GetAlbum(ctx context.Context, key string) (*photoslibrary.Album, error)
 
-func NewCachitaCache() *CachitaCache {
-	return &CachitaCache{cache: cachita.Memory()}
+	// PutAlbum stores the Album data in the cache under the specified key.
+	// Underlying implementations may use any data storage format,
+	// as long as the reverse operation, GetAlbum, results in the original data.
+	PutAlbum(ctx context.Context, key string, album *photoslibrary.Album, ttl time.Duration) error
+
+	// DeleteAlbum removes the Album data from the cache under the specified key.
+	// If there's no such key in the cache, DeleteAlbum returns nil.
+	InvalidateAlbum(ctx context.Context, key string) error
 }
